Added Japanese (kana) as a language
This is a list of Japanese words written in hiragana (as opposed to Kanji) making it easier to focus on typing on a kana keyboard. 

The words were taken from this page on Wiktionary: https://en.wiktionary.org/wiki/Appendix:1000_Japanese_basic_words)

They should all be "common use" or "basic" words.